Отрисовываются тени

+++ includes/draw.h
@@ -6,7 +6,7 @@
 /*   By: mcherrie <mcherrie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/01 12:06:29 by marlean           #+#    #+#             */
-/*   Updated: 2022/07/09 13:43:59 by mcherrie         ###   ########.fr       */
+/*   Updated: 2022/07/13 13:28:51 by mcherrie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,7 +63,7 @@ float cylindr_intersect(struct s_camera cam, t_coord ray, t_cylind *cylind);
 char *ft_dectohex(unsigned int num);
 unsigned long htoi(const char *s);
 
-int	draw_dot(t_data *data, t_dist *dist, int flag);
+int	draw_dot(t_data *data, t_dist *dist, float flag);
 
 //dist_obj.c
 void nearest_sphere(t_data *data, t_dist *dist, t_coord *ray);
